Initial management dummy api
Added new api module for communication with openstack-management
service. Structure of this module is similar to other existing
api modules (IOW plain list of api methods). True is this plain
structure of api methods is not optimal because there is ATM not good
place to put business logic, but this is horizon wide issue and will
be targeted in a separate patch.

Because there is not real management service yet, these api methods
use classic Django models for storing dummy data.

You have to run "python manage.py syncdb" to init/update db.

Change-Id: I4e56be6824061f55021819ff66faa1cfd7f135d2