Added additional flavor attributes to detail page
It looks like the database allows a flavor to be associated
with the same resource class multiple times.  This is a bug,
in our temporary API, but one that should go away when the
"real" API comes into play, so I don't see the need to fix
it now.

The instance count returns something arbitrary for now.

Change-Id: I1ec11627babe5056bfc22fa737e1cab222313f5e